Reworte the Parser, so it now only allow rows of same length as the header.
This has cuased the empty line bug to dissapear, since an empty line will now cause the parser to consider the csv file done. Except for cases where the header has length 1, where an empty line and an empty cell is really the same thing. Maybe the parser can be tweaked to be a little more tolerant, but I consider it more or less done.